Implement address connect and remove login (#3721)
* init

* Login from setting page button

* Update setting proxy page

* refactor

* refactor

* Update comment & post

* Update signer popup

* Update

* fix

* refactor

* refactor

* fix: login redirect

* fix

* fix: mail login redirect

* Update kintsugi

* fix

* refactor api

* fix

* fix

* Update text login to connect

* Update

* Extract settings page common (#3728)

* Extract settings page common

* Update

* Rename

* fix: web2 login

* fix email verify

* Remove hook useSetUserStatus (#3736)

* fix: event delete button

* Remove user in redux (#3739)

* Refactor login (#3742)

* Old login compatibility

* refactor login

* Update

* fix

* fix signer popup

* Update

* fix

* fix

* Simplify the code for account settings tabs (#3743)

* fix: key user check on setting pages (#3744)

* fix: key user check on setting pages

* rollback

* Rename

* fix

* fix link address

* fix: link metamask address

* refactor getAccountMenu to useAccountMenu (#3745)

* refactor getAccountMenu to useAccountMenu

* Add comment

* Extract settings common fields for account menu (#3746)

* fix: force to clean auth cookie when disconnect (#3747)

* Fix typo (#3748)

* Add loading effect to connect button (#3749)

* Reduce json stringify for account (#3750)

* Add comment

* fix thumb up loading icon (#3751)

* Update

* Add connect loading effect for singer popup wallet select (#3752)

* fix: address select filter

* refactor

* Update

---------

+++ packages/next-common/components/calendar/dayEvents/userEventInfoCard/deleteEventModal/index.js
@@ -2,16 +2,21 @@ import React, { useCallback, useState } from "react";
 import { useDispatch } from "react-redux";
 import nextApi from "next-common/services/nextApi";
 import Popup from "../../../../../components/popup/wrapper/Popup";
-import { newErrorToast, newSuccessToast } from "../../../../../store/reducers/toastSlice";
+import {
+  newErrorToast,
+  newSuccessToast,
+} from "../../../../../store/reducers/toastSlice";
 import { calendarUserEventsApi } from "../../../../../services/url";
 import ErrorMessage from "../../../../styled/errorMessage";
 import DangerButton from "../../../../buttons/dangerButton";
 import noop from "lodash.noop";
 import { PopupButtonWrapper } from "../../../../popup/wrapper";
+import { useEnsureLogin } from "next-common/hooks/useEnsureLogin";
 
 export default function DeleteEventModal({ event, onClose, refresh = noop }) {
   const dispatch = useDispatch();
   const [isLoading, setIsLoading] = useState(false);
+  const { ensureLogin } = useEnsureLogin();
 
   const showErrorToast = (message) => dispatch(newErrorToast(message));
   const showSuccessToast = (message) => dispatch(newSuccessToast(message));
@@ -19,6 +24,10 @@ export default function DeleteEventModal({ event, onClose, refresh = noop }) {
   const deleteEvent = useCallback(async () => {
     setIsLoading(true);
     try {
+      if (!(await ensureLogin())) {
+        return;
+      }
+
       const { result, error } = await nextApi.delete(
         `${calendarUserEventsApi}/${event?._id}`,
       );
@@ -35,7 +44,7 @@ export default function DeleteEventModal({ event, onClose, refresh = noop }) {
     } finally {
       setIsLoading(false);
     }
-  }, [dispatch, onClose, event?._id]);
+  }, [dispatch, onClose, event?._id, ensureLogin]);
 
   return (
     <Popup title="Delete" onClose={onClose}>
